Updated sql.

+++ tables/CountriesTable.sql
@@ -1,12 +1,12 @@
 CREATE TABLE IF NOT EXISTS `country`
 (
-    `id`        int(11)     NOT NULL AUTO_INCREMENT,
+    `id`        smallint    NOT NULL AUTO_INCREMENT,
     `iso`       char(2)     NOT NULL,
     `name`      varchar(80) NOT NULL,
     `nicename`  varchar(80) NOT NULL,
-    `iso3`      char(3)     DEFAULT NULL,
-    `numcode`   smallint(6) DEFAULT NULL,
-    `phonecode` int(5)      NOT NULL,
+    `iso3`      char(3)  DEFAULT NULL,
+    `numcode`   smallint DEFAULT NULL,
+    `phonecode` int         NOT NULL,
     PRIMARY KEY (`id`)
 );
